Add simple "fundamentally" implementation
It's not sure if this implementation does the same as the GTK's GObject.
Reason for this is, that the function is undocumented other than it's prototype.
This function is only required by us to later check, if we're going to initialize the refcount for a type provided to g_object_new.